Create mapping.json
This is for CodeReview. Modify your repo to contain a mapping.json file in the project root

This file will map your team's numerical GTIDs (903...) to your GitHub usernames. It is crucial that you type your GTID correctly and include your username, not your email
The file should be in the root (top folder) of your project